pricebot_portfolio/requirements.txt
- requirements.txt for Price Comparison Bot

requests
beautifulsoup4
pandas
openpyxl
python-dotenv
pytest  


smtplib
email
